requirements.txt
flask==1.0.2
python-dotenv==0.8.2
flask-sqlalchemy==2.3.2
psycopg2==2.7.4
flask-redis==0.3.0